Some fixes, need more
Created Exception.h, it's working. Need to fix resize.
